illicit-crypto-explorer-proto/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pandas



